Update based on feedback
Specifically:
* Updated some styling on imports
* Added a new keybindings module to represent keybindings
  for the entire application and moved keybindings from
  individual prompts to it
* Reference buffers by value name to simplify required reference
  values
